refactor: Error Handle 로직 이동.
- APIError 타입에 커스텀 이니셜라이저를 통한 처리.
- APIError 타입 별도의 파일로 분리.
